Move all old drawable note to `Graphics` folder
