+++ maisim/maisim.Game.Tests/Visual/Component/Gameplay/TestSceneMaisimRing.cs
@@ -1,4 +1,4 @@
-using maisim.Game.Component.Gameplay;
+using maisim.Game.Graphics.Gameplay;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osuTK;
