+++ maisim/maisim.Game.Tests/Visual/Screen/TestSceneNoteSpawnTimer.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq;
-using maisim.Game.Component.Gameplay.Notes;
+using maisim.Game.Gameplay.Notes;
+using maisim.Game.Graphics.Gameplay.Notes;
 using maisim.Game.Screen.Gameplay;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
