+++ maisim/maisim.Game/Graphics/Gameplay/Notes/DrawableSlideStarNote.cs
@@ -1,9 +1,8 @@
-﻿using maisim.Game.Screen.Gameplay;
-using osu.Framework.Graphics;
+﻿using osu.Framework.Graphics;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Textures;
 
-namespace maisim.Game.Component.Gameplay.Notes
+namespace maisim.Game.Graphics.Gameplay.Notes
 {
     /// <summary>
     /// Class represent the star in SLIDE note.
