+++ maisim/maisim.Game/Graphics/Gameplay/Notes/DrawableTapNote.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using maisim.Game.Gameplay.Notes;
 using maisim.Game.Screen.Gameplay;
 using maisim.Game.Utils;
 using osu.Framework.Allocation;
@@ -7,7 +8,7 @@ using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Textures;
 using osuTK;
 
-namespace maisim.Game.Component.Gameplay.Notes
+namespace maisim.Game.Graphics.Gameplay.Notes
 {
     /// <summary>
     /// Class represent drawable of the normal TAP note.
