+++ maisim/maisim.Game/maisim.Game.csproj
@@ -15,6 +15,6 @@
         <PackageReference Include="ppy.osu.Framework" Version="2022.816.0" />
     </ItemGroup>
     <ItemGroup>
-      <Folder Include="Component\Menu" />
+      <Folder Include="Gameplay" />
     </ItemGroup>
 </Project>
